@@ -10,7 +10,7 @@
     "lint": "ng lint",
     "reset": "rm -rf node_modules && rm -f package-lock.json && pnpm i",
     "build:serve": "ng build && serve dist/ng-b357/browser -s",
-    "serve": "serve dist/ng-b357/browser -s",
+    "serve": "npx serve dist/ng-b357/browser -s",
     "caddy": "caddy start",
     "caddy:stop": "caddy stop",
     "build:sme": "ng b --source-map && npm run sme",
@@ -22,27 +22,29 @@
     "wba": "npx webpack-bundle-analyzer dist/ng-b357/stats.json",
     "igv": "npx @rx-angular/import-graph-visualizer --entry-points src/main.ts --ts-config tsconfig.app.json",
     "dpdm": "npx dpdm --no-warning --no-tree src/main.ts",
-    "ng:update": "ng update @angular/cli @angular/core @angular/material angular-eslint"
+    "ng:update": "ng update @angular/cli @angular/core @angular/material angular-eslint",
+    "e2e": "ng e2e --ui"
   },
   "private": true,
   "dependencies": {
     "@angular-architects/paper-design": "^2.0.1",
-    "@angular/animations": "^19.2.5",
-    "@angular/cdk": "^19.2.8",
-    "@angular/common": "^19.2.5",
-    "@angular/compiler": "^19.2.5",
-    "@angular/core": "^19.2.5",
-    "@angular/forms": "^19.2.5",
-    "@angular/material": "^19.2.8",
-    "@angular/platform-browser": "^19.2.5",
-    "@angular/platform-browser-dynamic": "^19.2.5",
-    "@angular/platform-server": "^19.2.5",
-    "@angular/router": "^19.2.5",
-    "@angular/service-worker": "^19.2.5",
-    "@angular/ssr": "^19.2.6",
+    "@angular/animations": "^19.2.10",
+    "@angular/cdk": "^19.2.15",
+    "@angular/common": "^19.2.10",
+    "@angular/compiler": "^19.2.10",
+    "@angular/core": "^19.2.10",
+    "@angular/forms": "^19.2.10",
+    "@angular/material": "^19.2.15",
+    "@angular/platform-browser": "^19.2.10",
+    "@angular/platform-browser-dynamic": "^19.2.10",
+    "@angular/platform-server": "^19.2.10",
+    "@angular/router": "^19.2.10",
+    "@angular/service-worker": "^19.2.10",
+    "@angular/ssr": "^19.2.11",
+    "@axe-core/playwright": "^4.10.1",
     "@jsverse/transloco": "^7.6.1",
     "anychart": "^8.13.0",
-    "express": "^4.21.1",
+    "express": "^4.21.2",
     "moment": "^2.30.1",
     "ngx-quicklink": "^0.4.7",
     "rxjs": "^7.8.2",
@@ -50,26 +52,28 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.6",
-    "@angular/cli": "^19.2.6",
-    "@angular/compiler-cli": "^19.2.5",
+    "@angular-devkit/build-angular": "^19.2.11",
+    "@angular/cli": "^19.2.11",
+    "@angular/compiler-cli": "^19.2.10",
+    "@playwright/test": "1.52.0",
     "@types/express": "^5.0.1",
-    "@types/jasmine": "~5.1.7",
-    "@types/node": "^22.14.0",
-    "angular-eslint": "^19.3.0",
-    "eslint": "^9.24.0",
-    "eslint-config-prettier": "^10.1.1",
-    "eslint-plugin-prettier": "^5.2.6",
-    "jasmine-core": "~5.6.0",
+    "@types/jasmine": "~5.1.8",
+    "@types/node": "^22.15.17",
+    "angular-eslint": "^19.4.0",
+    "eslint": "^9.26.0",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-prettier": "^5.4.0",
+    "jasmine-core": "~5.7.1",
     "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
+    "playwright-ng-schematics": "2.0.3",
     "prettier": "^3.5.3",
     "source-map-explorer": "^2.5.3",
-    "typescript": "~5.7.3",
-    "typescript-eslint": "^8.29.0",
+    "typescript": "~5.8.3",
+    "typescript-eslint": "^8.32.0",
     "vite-bundle-visualizer": "^1.2.1",
     "webpack-bundle-analyzer": "^4.10.2"
   }
