@@ -1,4 +1,4 @@
-import { Component, ElementRef, EventEmitter, inject, Input, Output } from '@angular/core';
+import { Component, ElementRef, inject, model } from '@angular/core';
 
 import { BlinkService } from '../../shared/blink.service';
 
@@ -8,14 +8,13 @@ import { BlinkService } from '../../shared/blink.service';
   // changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class FlightStatusToggleComponent {
-  @Input({ required: true }) status = false;
-  @Output() readonly statusChange = new EventEmitter<boolean>();
+  readonly status = model.required<boolean>();
 
   private readonly blinkService = inject(BlinkService);
   private readonly elementRef = inject(ElementRef);
 
   protected onToggleStatus(): void {
-    this.statusChange.emit(!this.status);
+    this.status.set(!this.status());
   }
 
   protected blinkFirstChild(): void {
